servicios/servicio-pacientes/requirements.txt
fastapi>=0.119.1
uvicorn>=0.34.0
pydantic>=2.12.3
pymongo>=4.10.0
kafka-python>=2.0.2
python-dotenv>=1.0.1
py-eureka-client>=0.13.3